requirements.txt
flask==2.3.3
werkzeug==2.3.7
tensorflow==2.16.1
pillow==10.2.0
numpy==1.26.3
requests==2.31.0
gunicorn==21.2.0